微調節

+++ chatapp/views.py
@@ -18,11 +18,14 @@ def get_user(request, user_id):
 @csrf_exempt
 def message(request):
     if request.method == 'GET':
-        messages = ChatappMessage.objects.all()
+        # SQLを飛ばしている
+        # select * from ChatappMessage
+        messages = ChatappMessage.objects.all() # list[ChatappMessage]
         data = []
         for message in messages:
+            # message: ChatappMessage
             data.append({
-                'message': message.message,
+                'message': { "text": message.message, "id": message.id },
                 'sendername': message.sendername.name,
                 'sendername_id': message.sendername.id,
                 'created_at': [timezone.localtime(message.created_at).month,
@@ -36,12 +39,40 @@ def message(request):
         datas = json.loads(request.body)
         chatappUser_id = datas['chatappUser_id']
         message = datas['message']
+        # SQLを飛ばしている
+        # insert into ChatappMessage (sendername_id, message) values (chatappUser_id, message)
         ChatappMessage.objects.create(
             sendername=ChatappUser.objects.get(id=chatappUser_id), message=message)
         return HttpResponse('message登録完了！')
 
+
+@csrf_exempt
+def message_detail(request, message_id):
+    if request.method == 'GET':
+        message = ChatappMessage.objects.get(id=message_id)
+        data = {
+            'message': message.message,
+            'sendername': message.sendername.name,
+            'sendername_id': message.sendername.id,
+            'created_at': [timezone.localtime(message.created_at).month,
+                           timezone.localtime(message.created_at).day,
+                           timezone.localtime(message.created_at).hour,
+                           timezone.localtime(message.created_at).minute,]
+        }
+        return JsonResponse(data)
+
+    if request.method == 'PUT':
+        datas = json.loads(request.body)
+        chatappUser_id = datas['chatappUser_id']
+        message = datas['message']
+        chat_app_message = ChatappMessage.objects.get(id=message_id)
+        chat_app_message.sendername = ChatappUser.objects.get(id=chatappUser_id)
+        chat_app_message.message = message
+        chat_app_message.save()
+        return HttpResponse("更新に成功しました")
+
     if request.method == 'DELETE':
-        chat_app_message = ChatappMessage.objects.get(id=user_id)
+        chat_app_message = ChatappMessage.objects.get(id=message_id)
         chat_app_message.delete()
         return HttpResponse("削除に成功しました")
